Вынести расчёт точек из предложений в функцию
